Removed parametersetdb. Still used parts moved.
