+++ src/ewatercycle/plugins/lisflood/lisvap.py
@@ -27,7 +27,7 @@ from typing import Dict, Tuple
 
 from ewatercycle import CFG
 from ewatercycle.container import ContainerEngine
-from ewatercycle.parametersetdb.config import XmlConfig
+from ewatercycle.plugins.lisflood.config import XmlConfig
 from ewatercycle.plugins.lisflood._lisflood_versions import version_images
 from ewatercycle.util import get_time
