+++ src/ewatercycle/plugins/lisflood/model.py
@@ -13,7 +13,7 @@ from ewatercycle import CFG
 from ewatercycle.container import start_container
 from ewatercycle.base.model import AbstractModel
 from ewatercycle.base.parameter_set import ParameterSet
-from ewatercycle.parametersetdb.config import XmlConfig
+from ewatercycle.plugins.lisflood.config import XmlConfig
 from ewatercycle.plugins.lisflood._lisflood_versions import version_images
 from ewatercycle.plugins.lisflood.forcing import LisfloodForcing
 from ewatercycle.util import find_closest_point, get_time, to_absolute_path
