+++ tests/plugins/lisflood/test_lisvap.py
@@ -3,7 +3,9 @@ import subprocess
 from unittest.mock import patch
 
 from ewatercycle import CFG
-from ewatercycle.plugins.lisflood.lisvap import XmlConfig, create_lisvap_config, lisvap
+from ewatercycle.plugins.lisflood.lisvap import create_lisvap_config, lisvap
+
+from ewatercycle.plugins.lisflood.config import XmlConfig
 
 
 def find_values_in_xml(tree, name):
