+++ tests/plugins/lisflood/test_model.py
@@ -12,7 +12,7 @@ from numpy.testing import assert_array_equal
 from ewatercycle import CFG
 from ewatercycle.forcing import load_foreign
 from ewatercycle.base.parameter_set import ParameterSet
-from ewatercycle.parametersetdb.config import XmlConfig
+from ewatercycle.plugins.lisflood.config import XmlConfig
 from ewatercycle.plugins.lisflood.model import Lisflood
 from ewatercycle.testing.fake_models import FailingModel
