+++ tests/plugins/wflow/test_model.py
@@ -11,7 +11,7 @@ from grpc4bmi.bmi_client_apptainer import BmiClientApptainer
 
 from ewatercycle import CFG
 from ewatercycle.base.parameter_set import ParameterSet
-from ewatercycle.parametersetdb.config import CaseConfigParser
+from ewatercycle.util import CaseConfigParser
 from ewatercycle.plugins.wflow.model import Wflow
 from ewatercycle.testing.fake_models import FailingModel
